+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/). This project follows the conventions of [semver.org](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Corrected the identification of SQL Server's "INT IDENTITY" as a string.
+
 ### Added
 - Added the option to pass in a fixed list of tables from the command line rather than scanning the full database.
 - Added a command line option (`--ignore` / `-i`) to ignore a column during data generation.
 - Added a null data generation option to give the user a way to force a column to null.
+- Money / currency types to the list recognized by the decimal generator.
 
 ### Changed
 - Changed the long option for the output directory to `--directory` for `--output-dir`.